Fix issue where input is cleared after typing into it
In this test, the input was being typed into, then cleared before
the submit button was clicked. This resulted in a login failure.

By bypassing the type() method and directly setting the value of
the field, and by using setTimeout() to force the clock to tick
(and any ongoing events and animations to finish) before setting
the value, we can avoid synchronisation issues like this.